Split up integration tests and update chromedriver to support chrome 64
