+++ package.json
@@ -37,7 +37,7 @@
     "babel-preset-es2015": "^6.22.0",
     "babel-preset-react": "^6.22.0",
     "buffer-loader": "0.0.1",
-    "chromedriver": "2.34.1",
+    "chromedriver": "2.35.0",
     "classnames": "2.2.5",
     "copy-webpack-plugin": "^4.3.0",
     "css-loader": "^0.28.7",
