+++ test/helpers/selenium-helper.js
@@ -21,6 +21,17 @@ class SeleniumHelper {
         ]);
     }
 
+    get scope () {
+        // List of useful xpath scopes for finding elements
+        return {
+            blocksTab: "*[@id='react-tabs-1']",
+            modal: '*[@class="ReactModalPortal"]',
+            reportedValue: '*[@class="blocklyDropDownContent"]',
+            soundsTab: "*[@id='react-tabs-5']",
+            spriteTile: '*[starts-with(@class,"react-contextmenu-wrapper")]'
+        };
+    }
+
     getDriver () {
         this.driver = new webdriver.Builder()
             .forBrowser('chrome')
@@ -70,6 +81,12 @@ class SeleniumHelper {
     }
 
     getLogs (whitelist) {
+        if (!whitelist) {
+            // Default whitelist
+            whitelist = [
+                'The play() request was interrupted by a call to pause()'
+            ];
+        }
         return this.driver.manage()
             .logs()
             .get('browser')
