+++ test/integration/examples.test.js
@@ -13,10 +13,6 @@ const {
     loadUri
 } = new SeleniumHelper();
 
-const errorWhitelist = [
-    'The play() request was interrupted by a call to pause()'
-];
-
 let driver;
 
 describe('player example', () => {
@@ -37,7 +33,7 @@ describe('player example', () => {
         await clickXpath('//img[@title="Go"]');
         await new Promise(resolve => setTimeout(resolve, 2000));
         await clickXpath('//img[@title="Stop"]');
-        const logs = await getLogs(errorWhitelist);
+        const logs = await getLogs();
         await expect(logs).toEqual([]);
     });
 });
@@ -60,7 +56,7 @@ describe('blocks example', () => {
         await clickXpath('//img[@title="Go"]');
         await new Promise(resolve => setTimeout(resolve, 2000));
         await clickXpath('//img[@title="Stop"]');
-        const logs = await getLogs(errorWhitelist);
+        const logs = await getLogs();
         await expect(logs).toEqual([]);
     });
 
@@ -82,7 +78,7 @@ describe('blocks example', () => {
         el = await findByXpath("//input[@placeholder='']");
         await el.sendKeys('second variable');
         await clickButton('OK');
-        const logs = await getLogs(errorWhitelist);
+        const logs = await getLogs();
         await expect(logs).toEqual([]);
     });
 });
